requirements.txt
Django>=1.11.2,<2.0
# psycopg2 deprecated after version 2.8 use binary instead
psycopg2-binary
django-widget-tweaks>=1.4.1
selenium==3.6.0
rules==1.2.1
requests==2.18.4
django-environ==0.4.4
pillow==4.3.0
dj-database-url==0.4.2

# Test Coverage tools
coverage >= 4.4.1
django-nose>=1.2.1

# Linter tools
pycodestyle>=2.3.1
autopep8>=1.3.2

pylint-django
prospector

# GraphQL Dependencies
graphene>=2.0
graphene-django>=2.0
django-filter==1.1.0

# Nested Inlines
django-nested_admin==3.1.1